requirements.txt
groq==0.37.1
python-dotenv==1.2.1
langchain==0.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
pytest==7.4.3
